licenses/ and the first Generate scaffolding for a new project based on a template.
If only one argument is passed, the default template is used and the
argument is treated as if it were the name of the project.

Use ":show" instead of a project name to show template details.

Subtasks available:
default    A general project template.
plugin     A leiningen plugin project template.
template   A meta-template for 'lein new' templates.

Arguments: ([project-name] [template project-name & args]) w/ some deps spec'd